refactor: clarify IndexedDB compute ownership
